Apply clustering algorithms

Apply clustering algorithms

make_clusters(
  bio_df,
  env_df = NULL,
  methods_df = tibble(method = "average"),
  context,
  taxa_col = "taxa",
  num_col = NULL,
  env_cols,
  groups = 2:100,
  cores = 1,
  clust_col_in = "clust",
  clust_col_out = "cluster"
)

Arguments

bio_df
Dataframe. Needs the 'cols' outlined below.

env_df
Dataframe (wide) of environmental variables per sitecol.

methods_df
Dataframe. What methods to use to create clusters.

context
Character. Name of columns in bio_df that define the context.

taxa_col
Character. Name of column with taxa.

num_col
Character. Name of column with numeric data indicating relative abundance of taxa.

env_cols
Character. Name of columns in env_df that have the

groups
Numeric vector indicating the range of groups within a clustering.

cores
Numeric. Number of cores to use where parallel options exist.

clust_col_in
Character name of column of raw class labels.

clust_col_out
Character name of column of character class labels.

Value

Dataframe with groups and methods columns and a list column of the clustering for that number of groups and method as a tibble with one column (numeric) 'clust' indicating group membership.