Explore clusters

Deprecated. Use individual make_ functions.

make_clusters_explore(
  clusters_df,
  n_sample = 10,
  cores = 1,
  save_results = TRUE,
  do_env = TRUE,
  do_ind = TRUE,
  out_exp = tempdir(),
  exp_type = "cluster",
  do_gc = FALSE,
  obj_list
)

Arguments

clusters_df

Dataframe with column 'clusters'.

n_sample

Number of samples to use to calculate gap statistic.

cores

How many cores to use in multidplyr::new_cluster()?

save_results

Logical. If true, results will be saved along the way. If the file already exists, that result will not be recreated.

do_env

Logical. Run make env silhouette, wss and gap.

do_ind

Logical. Make

out_exp

Directory into which results are saved.

exp_type

Name to prefix files with and appended to out_exp.

do_gc

Run gc() after each output to save memory across cores.

obj_list

List of objects required:

dist_bio

dist(bio_wide).

dist_env

dist(env_wide).

bio_wide

Wide version of bio_tidy.

bio_tidy

Cleaned data set of taxa observations.

visit_cols

context.

p_thresh

usually 0.05.

n_sites

How many unique contexts in bio_tidy.

most_freq_prop_thresh

Threshold for proportion of sites in a cluster that must all contain at least one taxa in common.

Value

Each output saved along the way to out_exp. Outputs are not made if they have already been saved. Created outputs returned in list with elements:

clusters_sil

biota result from make_sil_df().

clusters_wss

biota result from calc_wss().

clusters_sil_env

env result from make_sil_df().

clusters_wss_env

env result from calc_wss().

clusters_ind_val

indicator values for biota from make_ind_val_df(), clusters_with_indicator() and sites_with_indicator().

clusters_freq

biota results from clusters_with_freq_taxa() and sites_with_freq_taxa()

Details

Explore clusters using silhouette width, within-group sum-of-squares, indicators and widespread taxa.