Explore clusters

Explore clusters using silhouette width, within-group sum-of-squares, indicators and widespread taxa.

make_clusters_explore(clusters_df, cores = 1, obj_list)

Arguments

clusters_df
Dataframe with column 'clusters'.

cores
How many cores to use in multidplyr::new_cluster()?

obj_list
List of objects required:

out_dir

directory into which to save results.

dist_flor

dist(flor_wide).

dist_flor_mat

as.matrix(dist(flor_wide)).

dist_env

dist(env_wide).

dist_env_mat

as.matrix(dist(env_wide)).

flor_wide

Wide version of flor_tidy.

flor_tidy

Cleaned data set of taxa observations.

visit_cols

context.

p_thresh

usually 0.05.

n_sites

How many unique contexts in flor_tidy.

most_freq_prop_thresh

Threshold for proportion of sites in a cluster that must all contain at least one taxa in common.

Value

clusters_df with added columns. Saves each output along the way to out_dir.